refactor: [#56] move upgrader from main upgrade mod to specific version upgrader mod
