+++ src/databases/mod.rs
@@ -1,5 +1,3 @@
 pub mod database;
 pub mod mysql;
 pub mod sqlite;
-pub mod sqlite_v1_0_0;
-pub mod sqlite_v2_0_0;
