+++ src/upgrades/from_v1_0_0_to_v2_0_0/databases/mod.rs
@@ -1,0 +1,2 @@
+pub mod sqlite_v1_0_0;
+pub mod sqlite_v2_0_0;
