+++ src/upgrades/from_v1_0_0_to_v2_0_0/databases/sqlite_v1_0_0.rs
@@ -1,7 +1,8 @@
-use super::database::DatabaseError;
 use serde::{Deserialize, Serialize};
 use sqlx::sqlite::SqlitePoolOptions;
 use sqlx::{query_as, SqlitePool};
+
+use crate::databases::database::DatabaseError;
 
 #[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
 pub struct Category {
@@ -22,9 +23,11 @@ impl SqliteDatabaseV1_0_0 {
     }
 
     pub async fn get_categories_order_by_id(&self) -> Result<Vec<Category>, DatabaseError> {
-        query_as::<_, Category>("SELECT category_id, name FROM torrust_categories ORDER BY category_id ASC")
-            .fetch_all(&self.pool)
-            .await
-            .map_err(|_| DatabaseError::Error)
+        query_as::<_, Category>(
+            "SELECT category_id, name FROM torrust_categories ORDER BY category_id ASC",
+        )
+        .fetch_all(&self.pool)
+        .await
+        .map_err(|_| DatabaseError::Error)
     }
 }
