+++ src/upgrades/mod.rs
@@ -1,0 +1,1 @@
+pub mod from_v1_0_0_to_v2_0_0;
